msinventory/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0
django-environ==0.11.2
